requirements.txt
aiohttp==3.7.4.post0
beautifulsoup4==4.10.0
cssselect==1.1.0
databases==0.5.2
discord-timers==0.1.0
discord.py==1.7.3
dotdict==0.1
lxml==4.6.3
PyNaCl==1.4.0
python-dateutil==2.8.2
PyYAML~=6.0
urllib3==1.26.6
Wand==0.6.7
yt-dlp
requests~=2.26.0
spotipy~=2.19.0
mutagen~=1.45.1
sqlalchemy~=1.4.26